Scrum 86 cart items analytics

+++ src/main/java/org/example/dto/wine/WineResponseDto.java
@@ -1,5 +1,6 @@
 package org.example.dto.wine;
 
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
@@ -11,10 +12,10 @@ import lombok.experimental.Accessors;
 public class WineResponseDto {
     private Long id;
     private String wineName;
-    private Double price;
+    private BigDecimal price;
     private String countryOfOrigin;
     private String wineType;
-    private Double popularityRating;
+    private BigDecimal popularityRating;
     private List<String> occasions;
     private String productImage;
 }
